Fix Else

+++ codebook/Else/BitHack.cpp
@@ -1,4 +1,3 @@
-long long next_perm(long long v) {
-  long long t = v | (v - 1);
-  return (t + 1) | (((~t & -~t) - 1) >> (__builtin_ctz(v) + 1));
-}
+ll next_perm(ll v) { ll t = v | (v - 1);
+  return (t + 1) |
+    (((~t & -~t) - 1) >> (__builtin_ctz(v) + 1)); }
